Money as float64: the rich could not be paid
float32 keeps ~7 significant digits wherever it sits on the number line,
so its absolute resolution degrades with magnitude. A tick's wage is
0.00039 gold. At a balance of 2,500 the smallest representable step is
0.00024, so a wage is 1.6 of them; at 10,000 it is 0.4 -- below half a
step, which means the addition does not happen at all.

A worker holding ten thousand gold was measured receiving NOTHING across
eighty simulated days while their employer's till emptied normally. The
error scales with the fortune, so the richest characters were the least
able to earn, and the drain tracked the hoards nobody could explain.

Gold, Wage, revenue, Prices, Treasury, Works, Vault and consignment Price
are now float64. Stock stays float32: grain is counted, not banked, and
the money/quantity split is written into the signatures that cross it --
transact(..., want float32, price float64), consign likewise.

102 type errors across sixteen files, resolved by reading each rather
than by pattern: an earlier attempt was abandoned when the automation
began widening variables holding meals.

Determinism, batch independence and one-sided conservation all hold.
Every world now walks a different path, since rounding changed in every
transaction, but the same seed still walks the same path twice -- which
is the property replay and measurement depend on. Every baseline measured
before this commit is void, 1,072/434 included.

MEASURED COST, with population held equal at 77 and the road and clinic
hot paths stubbed in BOTH arms so only the money types differ:

  7,998 ns   float32
  12,886 ns  float64          +61%
  13,556 ns  roads and clinics live, so those cost ~5% between them

Character is 224 bytes, from 200 this morning: kin links took it to 208,
the clinic to 216 via Skill[NumStructTypes], and money to 224.

THE LEAK IS NOT CLOSED. Precision was 36% of it; over twenty years the
unexplained loss fell from 1,645 to 1,048, still 2% of the money supply.
The remainder is a rule that does not balance, narrowed by audit to
stepBehaviour at 0.000126 per occasion over 87,084 occasions in twenty
days. work(), buyAndEat() and settleSale() are clean on inspection;
provision() and pan() are the untested candidates -- pan decrements
World.GoldOre in float32 while crediting float64, which the migration did
not reach because gold in the ground lives in worldgen.

A correct partial explanation is more dangerous than a wrong one. The
float32 mechanism was demonstrable, sufficient-looking, and only two
thirds of the answer, and I stopped looking when I could show the
arithmetic.

Recorded also at TestCharacterStaysCompact: a first attempt at the cost
split was contaminated because isolation stubs written as one-liners were
reformatted by gofmt, after which the scripted removal matched nothing
and silently left them in for two benchmarks and a full gate. Grep for
the marker before trusting a number that depended on removing it.

+++ internal/sim/business.go
@@ -118,8 +118,8 @@ func (s *State) foundBusinesses() {
 		if c.r == Food && s.FoodDays() < 60 {
 			ratio = 2 * ScarcitySignal // short granaries outrank every other signal
 		}
-		if ratio > bestRatio {
-			bestType, bestRatio = c.t, ratio
+		if ratio > float64(bestRatio) {
+			bestType, bestRatio = c.t, float32(ratio)
 		}
 	}
 	if bestType == NumStructTypes {
@@ -169,10 +169,10 @@ const HarvestMargin = 0.6
 func (s *State) foundOneNear(t StructType, near torus.Cell) bool {
 	// The founder: whoever can best afford it. Ties break toward the lower ID so the
 	// choice never depends on iteration order (§9.2).
-	cost := s.materialCost(t) + s.labourCost(t)
+	cost := float64(s.materialCost(t)) + s.labourCost(t)
 	// The founder keeps back four households' reserves — nobody sinks their last coin
 	// into a venture.
-	founder, founderGold := NoChar, cost+float32(LarderReserve*4)
+	founder, founderGold := NoChar, cost+float64(LarderReserve*4)
 	for i := range s.Chars {
 		c := &s.Chars[i]
 		if c.Alive && c.Stage() != Child && c.Gold > founderGold {
